requirements.txt
certifi==2020.12.5
djitellopy2==2.2
numpy==1.20.1
opencv-python==4.5.1.48
wincertstore==0.2
